Modify the move function to handle snake growth
Added a ternary operator to the move function to replace the
current food element with a new one if it shares coordinates with the
head of the snake, or remove the last part of the snake otherwise.
Since a part is always added at the front of the snake everytime it
moves, this function will cause the snake to grow by one part if the
first condition is met because no part will be removed.