Remove wall collisions
Added a function that allows the snake to come out of the opposite
side of the board if it goes out of bounds. It returns the coordinates
of the opposite side of the board if the previous move made the snake
go out of bounds, otherwise it returns the expected the coordinates.
Also made a few tweaks in the move function to account for these
changes.